Example of using functions within functions
This exercise assumes the math() module didn't exist, so no min()
or max() functions, we have to create our own. It uses passing
function returns between functions, calling one function from
another to create a result. There is perhaps a better way I don't
know of yet, but it works and is understandable.