Experiment in creating a database from scratch
Created a database and some purposed files for working with it.
The work.db file is something of a gateway to the data, prompting
the user whether they want to add data or search it. Depending on
the answer, it calls one of the other files. It's all very basic
and rough, but working.